extract screens into their own classes, use polymorphism instead of switches
