+++ UI/app.py
@@ -1,0 +1,4 @@
+class ScreenManager:
+    def __init__(self):
+        self.screens = []
+        self.current_screen = None
